+++ config/view_helper.config.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Sake
+ *
+ * @link      http://github.com/sandrokeil/BlockchainWalletApi for the canonical source repository
+ * @copyright Copyright (c) 2014 Sandro Keil
+ * @license   http://github.com/sandrokeil/BlockchainWalletApi/blob/master/LICENSE.txt New BSD License
+ */
+
+return array(
+    'invokables' => array(
+       'satoshi' => '\Sake\BlockchainWalletApi\View\Helper\Satoshi',
+    ),
+);