+++ src/Module.php
@@ -10,11 +10,12 @@
 namespace Sake\BlockchainWalletApi;
 
 use Zend\ModuleManager\Feature\ServiceProviderInterface;
+use Zend\ModuleManager\Feature\ViewHelperProviderInterface;
 
 /**
  * This class initializes the BlockchainWalletApi module.
  */
-class Module implements ServiceProviderInterface
+class Module implements ServiceProviderInterface, ViewHelperProviderInterface
 {
     /**
      * Return service config
@@ -25,5 +26,16 @@ class Module implements ServiceProviderInterface
     public function getServiceConfig()
     {
         return require dirname(__DIR__) . '/config/service_manager.config.php';
+    }
+
+    /**
+     * Expected to return \Zend\ServiceManager\Config object or array to
+     * seed such an object.
+     *
+     * @return array
+     */
+    public function getViewHelperConfig()
+    {
+        return require dirname(__DIR__) . '/config/view_helper.config.php';
     }
 }