+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 1.0.0 - TBA
 
+* Satoshi view helper which converts satoshi to other unit e.g. bitcoin
 * Validator support via input filter
 * Service plugin manager for responses, requests and input filter
 * Hydrator strategies for address list and addresses